% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inch.R
\name{inchtocm}
\alias{inchtocm}
\title{Converting inches to centimetres}
\usage{
inchtocm(i)
}
\arguments{
\item{i}{The vector, in inches, to be converted to centimetres}
}
\value{
A vector, in centimetres
}
\description{
This function converts inches to centimetres
}
\details{
This function is built onto the `powers` package, which can be found here: \link{powers}
}
\examples{
inchtocm(1)
inchtocm(12)
inchtocm(3.9816)
}
